+++ src/wildboar/distance/_matrix_profile.py
@@ -3,16 +3,16 @@ import numbers
 import warnings
 
 import numpy as np
-from sklearn.utils.validation import check_scalar
+from sklearn.utils._param_validation import Interval, StrOptions, validate_params
 
-from ..utils.validation import _check_ts_array, check_array, check_type
+from ..utils.validation import _check_ts_array, check_array
 from . import _cmatrix_profile
 from ._distance import _format_return
 
 
 def paired_matrix_profile(  # noqa: PLR0912
-    x,
-    y=None,
+    X,
+    Y=None,
     *,
     window=5,
     dim=0,
@@ -23,25 +23,25 @@ def paired_matrix_profile(  # noqa: PLR0912
     """
     Compute the matrix profile.
 
-    - If only ``x`` is given, compute the similarity self-join of every subsequence in
-      ``x`` of size ``window`` to its nearest neighbor in `x` excluding trivial matches
-      according to the ``exclude`` parameter.
-    - If both ``x`` and ``y`` are given, compute the similarity join of every
-      subsequenec in ``y`` of size ``window`` to its nearest neighbor in ``x`` excluding
-      matches according to the ``exclude`` parameter.
+    - If only `X` is given, compute the similarity self-join of every subsequence in
+      `X` of size ``window`` to its nearest neighbor in `X` excluding trivial matches
+      according to the `exclude` parameter.
+    - If both `X` and `Y` are given, compute the similarity join of every
+      subsequenec in `X` of size `window` to its nearest neighbor in `Y` excluding
+      matches according to the `exclude` parameter.
 
     Parameters
     ----------
-    x : array-like of shape (n_timestep, ), (n_samples, xn_timestep) or \
-        (n_samples, n_dim, xn_timestep)
+    X : array-like of shape (n_timestep, ), (n_samples, x_timestep) or \
+        (n_samples, n_dim, x_timestep)
         The first time series.
-    y : array-like of shape (n_timestep, ), (n_samples, yn_timestep) or \
-        (n_samples, n_dim, yn_timestep), optional
-        The optional second time series. y is broadcast to the shape of x if possible.
+    Y : array-like of shape (n_timestep, ), (n_samples, y_timestep) or \
+        (n_samples, n_dim, y_timestep), optional
+        The optional second time series. Y is broadcast to the shape of X if possible.
     window : int or float, optional
         The subsequence size, by default 5
 
-        - if float, a fraction of `y.shape[-1]`.
+        - if float, a fraction of y_timestep
         - if int, the exact subsequence size.
     dim : int, optional
         The dim to compute the matrix profile for, by default 0.
@@ -65,10 +65,7 @@ def paired_matrix_profile(  # noqa: PLR0912
 
     Notes
     -----
-    The `profile_size` depends on the input.
-
-    - If `y` is `None`, `profile_size` is  ``x.shape[-1] - window + 1``
-    - If `y` is not `None`, `profile_size` is ``y.shape[-1] - window + 1``
+    The `profile_size` is  ``X.shape[-1] - window + 1``.
 
     References
     ----------
@@ -77,67 +74,47 @@ def paired_matrix_profile(  # noqa: PLR0912
         that includes motifs, discords and shapelets. In 2016 IEEE 16th international
         conference on data mining (ICDM)
     """
-    x = check_array(x, allow_3d=True, ensure_2d=False)
+    X = check_array(X, allow_3d=True, ensure_2d=False)
 
-    if y is not None:
-        y = check_array(y, allow_3d=True, ensure_2d=False)
-        if x.ndim > 1:
-            y = np.broadcast_to(y, x.shape)
+    if Y is not None:
+        Y = check_array(Y, allow_3d=True, ensure_2d=False)
+        if X.ndim > 1:
+            Y = np.broadcast_to(Y, X.shape)
 
-        if x.ndim != y.ndim:
+        if X.ndim != Y.ndim:
             raise ValueError("Both x and y must have the same dimensionality")
-        if x.ndim != 1 and x.shape[0] != y.shape[0]:
+        if X.ndim != 1 and X.shape[0] != Y.shape[0]:
             raise ValueError("Both x and y must have the same number of samples")
-        if x.ndim > 2 and x.shape[1] != y.shape[1]:
+        if X.ndim > 2 and X.shape[1] != Y.shape[1]:
             raise ValueError("Both x and y must have the same number of dimensions")
-        if not y.shape[-1] <= x.shape[-1]:
+        if not Y.shape[-1] <= X.shape[-1]:
             raise ValueError(
                 "y.shape[-1] > x.shape[-1]. If you want to compute the matrix profile "
                 "of the similarity join of YX, swap the order of inputs."
             )
         exclude = exclude if exclude is not None else 0.0
     else:
-        y = x
+        Y = X
         exclude = exclude if exclude is not None else 0.2
 
-    if x.ndim > 2 and not 0 <= dim < x.shape[1]:
-        raise ValueError("Invalid dim (%d)" % x.shape[1])
+    if not isinstance(window, numbers.Integral):
+        window = math.ceil(window * Y.shape[-1])
 
-    check_type(window, "window", (numbers.Integral, numbers.Real))
-    check_type(exclude, "exclude", (numbers.Integral, numbers.Real))
-    if isinstance(window, numbers.Integral):
-        check_scalar(
-            window,
-            "window",
-            numbers.Integral,
-            min_val=1,
-            max_val=min(y.shape[-1], x.shape[-1]),
-        )
-    elif isinstance(window, numbers.Real):
-        check_scalar(
-            window,
-            "window",
-            numbers.Real,
-            min_val=0,
-            max_val=1,
-            include_boundaries="right",
-        )
-        window = math.ceil(window * y.shape[-1])
+    if not (0 < window <= min(X.shape[-1], Y.shape[-1])):
+        raise ValueError("window must be 0 < window <= min(x_timestep, y_timestep)")
 
-    if isinstance(exclude, numbers.Integral):
-        check_scalar(exclude, "exclude", numbers.Integral, min_val=1)
-    elif isinstance(exclude, numbers.Real):
-        check_scalar(
-            exclude,
-            "exclude",
-            numbers.Real,
-            min_val=0,
-        )
+    if not isinstance(exclude, numbers.Integral):
         exclude = math.ceil(window * exclude)
 
+    if X.ndim > 2 and X.ndim != Y.ndim and X.shape[1] != Y.shape[1]:
+        raise ValueError("Both X and Y must have the same number of dimensions")
+
+    if X.ndim > 2 and not 0 <= dim < X.shape[1]:
+        raise ValueError("dim must be less than n_dims")
+
     mp, mpi = _cmatrix_profile._paired_matrix_profile(
-        _check_ts_array(x),
-        _check_ts_array(y),
+        _check_ts_array(Y),
+        _check_ts_array(X),
         window,
         dim,
         exclude,
@@ -145,12 +122,42 @@ def paired_matrix_profile(  # noqa: PLR0912
     )
 
     if return_index:
-        return _format_return(mp, 2, x.ndim), _format_return(mpi, 2, x.ndim)
+        return _format_return(mp, 2, X.ndim), _format_return(mpi, 2, X.ndim)
     else:
-        return _format_return(mp, 2, x.ndim)
+        return _format_return(mp, 2, X.ndim)
 
 
-def matrix_profile(X, Y=None, window=5, exclude=None, return_index=False, n_jobs=None):
+@validate_params(
+    {
+        "x": ["array-like"],
+        "y": [None, "array-like"],
+        "dim": [Interval(numbers.Integral, 0, None, closed="left")],
+        "window": [
+            Interval(numbers.Integral, 1, None, closed="left"),
+            Interval(numbers.Real, 0, 1, closed="both"),
+        ],
+        "exclude": [
+            None,
+            Interval(numbers.Integral, 1, None, closed="left"),
+            Interval(numbers.Real, 0, 1, closed="both"),
+        ],
+        "kind": [StrOptions({"default", "paired", "warn"}, deprecated={"warn"})],
+        "return_index": [bool],
+        "n_jobs": [None, numbers.Integral],
+    },
+    prefer_skip_nested_validation=True,
+)
+def matrix_profile(  # noqa: PLR0912
+    X,
+    Y=None,
+    *,
+    dim=0,
+    window=5,
+    exclude=None,
+    kind="warn",
+    return_index=False,
+    n_jobs=None,
+):
     """
     Compute the matrix profile of every subsequence in X.
 
@@ -170,6 +177,8 @@ def matrix_profile(X, Y=None, window=5, exclude=None, return_index=False, n_jobs
         The time series for which the matrix profile is computed.
     Y : array-like of shape (y_samples, y_timestep), optional
         The time series used to annotate `X`. If None, `X` is used to annotate.
+    dim : int, optional
+        The dimension.
     window : int or float, optional
         The window size.
 
@@ -183,6 +192,14 @@ def matrix_profile(X, Y=None, window=5, exclude=None, return_index=False, n_jobs
         - If None, the exclusion zone is determined automatically. If Y is None,
           (self-join) the value is `0.2`, otherwise (AB-join) the value is
           `0.0`.
+    kind : {"paired", "default"}, optional
+        The kind of matrix profile.
+
+        - if "paired", compute the matrix profile for each time series in X
+          optionally annotated with each time series in Y.
+        - if "default", compute the matrix profile for every subsequence in
+          every time series in X optional annotated with every time series in
+          Y.
     return_index : bool, optional
         Return the matrix profile index.
     n_jobs : int, optional
@@ -196,40 +213,54 @@ def matrix_profile(X, Y=None, window=5, exclude=None, return_index=False, n_jobs
         The matrix profile index sample and start positions. Returned if
         `return_index=True`.
     """
-    # TODO(1.4)
-    warnings.warn(
-        "matrix_profile has changed name to paired_matrix_profile in 1.3. "
-        "matrix_profile now computes the matrix profile for every subsequence in "
-        "all time series in X. This warning will be removed in 1.4.",
-        FutureWarning,
-    )
-    if Y is None:
-        Y = X
+    if kind in ("warn", "paired"):
+        if kind == "warn":
+            # TODO(1.4)
+            warnings.warn(
+                "matrix_profile has changed name to paired_matrix_profile in 1.3 "
+                "and the default behavior will change in 1.4",
+                FutureWarning,
+            )
+            if Y is not None:
+                X, Y = Y, X
 
-    if X is Y:
-        X = check_array(X, ensure_ts_array=True)
-        Y = X
-        if exclude is None:
-            exclude = 0.2
+        return paired_matrix_profile(
+            X, Y, window=window, dim=dim, exclude=exclude, n_jobs=n_jobs
+        )
     else:
-        X = check_array(X, ensure_ts_array=True)
-        Y = check_array(Y, ensure_ts_array=True)
-        if exclude is None:
-            exclude = 0.0
+        if Y is None:
+            Y = X
 
-    if not isinstance(window, numbers.Integral):
-        window = math.ceil(window * Y.shape[-1])
+        if X is Y:
+            X = check_array(X, ensure_ts_array=True)
+            Y = X
+            if exclude is None:
+                exclude = 0.2
+        else:
+            X = check_array(X, ensure_ts_array=True)
+            Y = check_array(Y, ensure_ts_array=True)
+            if exclude is None:
+                exclude = 0.0
 
-    if not (0 < window <= min(X.shape[-1], Y.shape[-1])):
-        raise ValueError("window must be 0 < window <= min(x_timestep, y_timestep)")
+        if not isinstance(window, numbers.Integral):
+            window = math.ceil(window * Y.shape[-1])
 
-    if not isinstance(exclude, numbers.Integral):
-        exclude = math.ceil(window * exclude)
+        if not (0 < window <= min(X.shape[-1], Y.shape[-1])):
+            raise ValueError("window must be 0 < window <= min(x_timestep, y_timestep)")
 
-    mp, mpi = _cmatrix_profile._flat_matrix_profile_join(
-        X, Y, window, 0, exclude, n_jobs
-    )
-    if return_index:
-        return mp, (mpi[:, :, 0], mpi[:, :, 1])
-    else:
-        return mp
+        if not isinstance(exclude, numbers.Integral):
+            exclude = math.ceil(window * exclude)
+
+        if X.shape[1] != Y.shape[1]:
+            raise ValueError("Both X and Y must have the same number of dimensions")
+
+        if not 0 <= dim < X.shape[1]:
+            raise ValueError("dim must be less than n_dims")
+
+        mp, mpi = _cmatrix_profile._flat_matrix_profile_join(
+            X, Y, window, dim, exclude, n_jobs
+        )
+        if return_index:
+            return mp, (mpi[:, :, 0], mpi[:, :, 1])
+        else:
+            return mp